Support for HTML5 compliant form errors ( was https://bitbucket.org/formencode/official-formencode/pull-request/16/support-removing-tags-makes-formencode ) * supports non-selfclosing form:error tags "<form:error name='a'></form:error>" * tests added to ensure backwards compliance of 'classic' self-closing tags , and new non-selfclosing tags. * tests also check to ensure the following tag is not affected
Recap:
On the html5 spec, only a subset of tags - void , svg, and math - are allowed to be self-closing. The rest are expected to close. [ http://www.w3.org/TR/html5/syntax.html#void-elements ]

The issue is that (according to the specs) while <form:error /> is valid for XHTML it is invalid for HTML5 . <form:error></form:error> should work on both.

This is only an issue if you're not parsing the original state of the form with htmlfill , or if you're having a front-end team mark up standalone html pages before passing off to the backend -- but it can pop up.