Added option to skip password rendering.
This feature is useful for websites that do not wish to return forms with
pre-filled passwords, which might be appropriate in security-conscious websites
and applications.

A small user experience consideration, though, make sure you do as much
error-handling as possible before submitting forms, to prevent having a user
fill out their password over and over again.