Adding self to CONTRIBUTORS.txt.
Tests were failing on formencode/national.py with AttributeErrors when integrating with pycountry. I can't find where the interface in pycountry changed, but it appears all instances of alpha2 should now be alpha_2.

I don't know the protocol, if I should fix this in this branch - it's totally unrelated - but I want the tests to pass.