Fix pkg_resources DeprecationWarning
pkg_resources now warns on absolute paths in resource_filename. Because
of how resource_filename handles this specific case (discarding the
leading '/' because of how it uses split and os.path.joini to construct
the actual name), the leading '/' here doesn't actually do anything, so
we can remove it without consequence.