Don't allow FieldStorageUploadConverter without cgi module
In Python 3.13 and later the cgi module has been removed and this meant
that importing the `cgi` module started making it so
`formencode.validators` could not be imported. This was resolved in #176
by detecting the import error, setting `cgi` to `None`, and then
skipping behaviour that used `cgi.FieldStorage`. Using
`FieldStorageUploadConverter` without `cgi.FieldStorage` does not make
sense. This change adds a check to
`FieldStorageUploadConverter.__init_` to prevent instantiating it
like we do for DNS validators so that using it doesn't silently do
nothing.

Probably if you are using FieldStorageUploadConverter then you'll have
`legacy-cgi` installed because you are either already using it directly,
or you are getting it transitively through something like WebOb.